pyproject.toml
[project]
name = "relaybot"
version = "0.1.0"
description = "northstar relay bot"
authors = [{ name = "Bobbyperson", email = "bobbyperson12345@gmail.com" }]
requires-python = ">=3.11,<3.13,!=3.12.5"
readme = "README.md"
license = "GNU GPLv3"
dependencies = [
    "discord.py[voice]>=2.5.2",
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "requests>=2.32.3",
    "aiohttp>=3.11.16",
    "python-dateutil>=2.9.0.post0",
    "keyring>=25.6.0",
    "discord-pretty-help>=2.0.7",
    "rcon>=2.4.9",
    "humanize>=4.12.2",
    "pillow>=11.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.4",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
